split up authentication from header && added publish build
