+++ components/ws-header/ws-header-nav-link.js
@@ -1,0 +1,9 @@
+import { React, Component } from '../imports';
+
+export default class WSHeaderNavLink extends Component {
+  render() {
+    return (<li className="nav-link" key={this.props.key} onClick={() => this.props.link.onclick && this.props.link.onclick()} >
+      <a>{this.props.link.label}</a>
+    </li>);
+  }
+}
