+++ demo/demo_app.js
@@ -1,6 +1,10 @@
 import { React, render } from '../components/imports';
-import WSHeader from '../components/ws-header.js';
+import WSHeader from '../components/ws-header/ws-header';
 
-render(<WSHeader title='Demo Page' links={[
-	{ label: 'Link', value: 'LinkValue', onclick: (value) => router.goTo(value) }
-]}/>, document.body);
+let linkValue = 'link_value';
+render(<WSHeader
+  title="Demo Page"
+  links={[
+    { label: 'Link', onclick: () => alert('Clicked on Link:', linkValue) },
+  ]}
+/>, document.body);
