+++ webpack.config.js
@@ -1,4 +1,4 @@
-var path = require('path');
+const path = require('path');
 
 module.exports = {
   entry: './demo/demo_app.js',
